+++ lua/mini/files.lua
@@ -657,9 +657,10 @@ end
 --- There is no constraint by default.
 ---
 --- `windows.preview` is a boolean indicating whether to show preview of
---- file/directory under cursor. Note: it is shown with highlighting if Neovim
---- version is sufficient and file is small enough (less than 1K bytes per line
---- or 1M bytes in total).
+--- file/directory under cursor. Notes:
+--- - It is always shown, even if current line is for not yet existing path.
+--- - File preview is highlighted if its size is small enough (less than 1K
+---   bytes per line or 1M bytes in total).
 ---
 --- `windows.width_focus` and `windows.width_nofocus` are number of columns used
 --- as `width` for focused and non-focused windows respectively.
@@ -1574,7 +1575,9 @@ H.explorer_sync_cursor_and_branch = function(explorer, depth)
   local cursor_path
   if type(cursor) == 'table' and H.is_valid_buf(buf_id) then
     local l = H.get_bufline(buf_id, cursor[1])
-    cursor_path = H.path_index[H.match_line_path_id(l)]
+    -- Fall back to treating current line as full basename, but for not yet
+    -- existing (a.k.a. "imaginary") path keep showing preview.
+    cursor_path = H.path_index[H.match_line_path_id(l)] or H.fs_child_path(path, l .. '\000')
   elseif type(cursor) == 'string' then
     cursor_path = H.fs_child_path(path, cursor)
   else
@@ -1591,9 +1594,8 @@ H.explorer_sync_cursor_and_branch = function(explorer, depth)
 
   -- Show preview to the right of current buffer if needed
   local show_preview = explorer.opts.windows.preview
-  local path_is_present = type(cursor_path) == 'string' and H.fs_is_present_path(cursor_path)
   local is_cur_buf = explorer.depth_focus == depth
-  if show_preview and path_is_present and is_cur_buf then table.insert(explorer.branch, cursor_path) end
+  if show_preview and is_cur_buf then table.insert(explorer.branch, cursor_path) end
 
   return explorer
 end
@@ -2092,6 +2094,12 @@ H.buffer_create = function(path, mappings)
   -- Register buffer
   H.opened_buffers[buf_id] = { path = path }
 
+  -- Set buffer options
+  vim.bo[buf_id].filetype = 'minifiles'
+
+  -- Do not set up tracking behavior for imaginary paths
+  if H.fs_is_imaginary_path(path) then return buf_id end
+
   -- Make buffer mappings
   H.buffer_make_mappings(buf_id, mappings)
 
@@ -2101,14 +2109,11 @@ H.buffer_create = function(path, mappings)
     vim.api.nvim_create_autocmd(events, { group = augroup, buffer = buf_id, desc = desc, callback = callback })
   end
 
-  au({ 'CursorMoved', 'CursorMovedI' }, 'Tweak cursor position', H.view_track_cursor)
+  au({ 'CursorMoved', 'CursorMovedI', 'TextChangedP' }, 'Tweak cursor position', H.view_track_cursor)
   au({ 'TextChanged', 'TextChangedI', 'TextChangedP' }, 'Track buffer modification', H.view_track_text_change)
 
   -- Tweak buffer to be used nicely with other 'mini.nvim' modules
   vim.b[buf_id].minicursorword_disable = true
-
-  -- Set buffer options
-  vim.bo[buf_id].filetype = 'minifiles'
 
   -- Trigger dedicated event
   H.trigger_event('MiniFilesBufferCreate', { buf_id = buf_id })
@@ -2206,14 +2211,17 @@ H.buffer_make_mappings = function(buf_id, mappings)
 end
 
 H.buffer_update = function(buf_id, path, opts, is_preview)
-  if not (H.is_valid_buf(buf_id) and H.fs_is_present_path(path)) then return end
+  if not H.is_valid_buf(buf_id) then return end
 
   -- Perform entry type specific updates
-  local update_fun = H.fs_get_type(path) == 'directory' and H.buffer_update_directory or H.buffer_update_file
-  update_fun(buf_id, path, opts, is_preview)
+  local fs_type = H.fs_get_type(path)
+  if fs_type == 'directory' then H.buffer_update_directory(buf_id, path, opts, is_preview) end
+  if fs_type == 'file' then H.buffer_update_file(buf_id, path, opts, is_preview) end
+  if fs_type == nil then H.set_buflines(buf_id, { '-No-fs-entry-' .. string.rep('-', opts.windows.width_preview) }) end
 
   -- Trigger dedicated event
-  H.trigger_event('MiniFilesBufferUpdate', { buf_id = buf_id, win_id = H.opened_buffers[buf_id].win_id })
+  local data = { buf_id = buf_id, win_id = H.opened_buffers[buf_id].win_id }
+  if fs_type ~= nil then H.trigger_event('MiniFilesBufferUpdate', data) end
 
   -- Reset buffer as not modified
   H.opened_buffers[buf_id].n_modified = -1
@@ -2591,7 +2599,9 @@ if H.is_windows then
   H.fs_normalize_path = function(path) return (path:gsub('\\', '/'):gsub('/+', '/'):gsub('(.)[\\/]$', '%1')) end
 end
 
-H.fs_is_present_path = function(path) return vim.loop.fs_stat(path) ~= nil end
+H.fs_is_imaginary_path = function(path) return path:sub(-1) == '\000' end
+
+H.fs_is_present_path = function(path) return vim.loop.fs_stat(path) ~= nil and not H.fs_is_imaginary_path(path) end
 
 H.fs_child_path = function(dir, name) return H.fs_normalize_path(string.format('%s/%s', dir, name)) end
 
@@ -2623,7 +2633,7 @@ end
 H.fs_is_windows_top = function(path) return H.is_windows and path:find('^%w:[\\/]?$') ~= nil end
 
 H.fs_get_type = function(path)
-  if not H.fs_is_present_path(path) then return nil end
+  if not (not H.fs_is_imaginary_path(path) and H.fs_is_present_path(path)) then return nil end
   return vim.fn.isdirectory(path) == 1 and 'directory' or 'file'
 end
 
@@ -2907,7 +2917,7 @@ H.getcharstr = function()
   return char
 end
 
-H.sanitize_string = function(x) return ((x or ''):gsub('\n', '<NL>')) end
+H.sanitize_string = function(x) return ((x or ''):gsub('\n', '<NL>'):gsub('%z', '')) end
 
 -- TODO: Remove after compatibility with Neovim=0.9 is dropped
 H.islist = vim.fn.has('nvim-0.10') == 1 and vim.islist or vim.tbl_islist